# public/docs/swagger.yaml
openapi: 3.0.0
info:
  title: MaryClean API REST
  description: >
    Documentación interactiva de la API de Lavandería MaryClean.
    
    ### Instrucciones de Autenticación 🔒
    1. Ejecuta el endpoint `POST /api/v1/auth/login`.
    2. Copia el token de la respuesta (`data.token`).
    3. Haz clic en el botón **Authorize** 🔒 (arriba a la derecha) e ingresa el token.
  version: 1.0.0
servers:
  - url: /
    description: Servidor Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/clientes/{id}:
    get:
      tags: [Clientes]
      summary: Obtener detalle de cliente
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Detalle del cliente
    put:
      tags: [Clientes]
      summary: Actualizar cliente
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                telefono: { type: string, example: "999888777" }
      responses:
        '200':
          description: Cliente actualizado

  # ----------------------------------------------------------------------
  # PEDIDOS
  # ----------------------------------------------------------------------
...
  /api/v1/pedidos/{id}:
    get:
      tags: [Pedidos]
      summary: Obtener pedido completo con detalles
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Pedido completo
...
  /api/v1/pedidos/{id}/estado:
    patch:
      tags: [Pedidos]
      summary: Cambiar estado del pedido
      description: Dispara auditoría MySQL. (Solo admin puede Cancelar)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                estado: { type: string, example: "En Proceso" }
      responses:
        '200':
          description: Estado actualizado
        '403':
          description: Rol insuficiente para cancelar

  # ----------------------------------------------------------------------
  # PAGOS
  # ----------------------------------------------------------------------
...
  /api/v1/pagos/pedido/{idPedido}:
    get:
      tags: [Pagos]
      summary: Historial de pagos de un pedido y saldo restante
      parameters:
        - in: path
          name: idPedido
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Historial y saldos
...